Switch to git commit hash instead of branch
The issue with using a tag is that on a redeploy, k8s will not see the
deployment as different, even though what the tag points to has changed.
By using the commit sha that is automatically added as a tag to the
image during the GitHub Actions process, k8s will always pick up the
right image.

Further automation is coming to help in this area as well.