conf(web): Add uwsgi config as ConfigMap
The original configuration is part of the base image, this doesn't allow
for dynamic modification and the default configuration is using a lot of
resources.

Use the default configuration as the source for the configuration and
adjusted the number of workers to 5, this number too may be too much,
but rather than add more workers, add another pod.

To create the ConfigMap, use the following command:

kubectl \
  create configmap uwsgi \
  -n apps--web \
  --from-file=uwsgi.ini \
  -o yaml \
  --dry-run=client \
  > configmap-uwsgi.yaml