feat(prometheus): Add prometheus for metrics collection
Initial commit of prometheus that collects and displays metrics, stores
them in Thanos, and includes dashboards in Grafana.

The installation uses the kube-prometheus project
(https://github.com/prometheus-operator/kube-prometheus), manifests
generation uses the `Makefile`, with application using kustomize.

As the upstream project using jsonnet for configuration, most
configuration options are in the settings.jsonnet file. Changes to the
settings, requires rerunning the `make manifests` to create the new
manifests. Commit the results to the metacpan-k8s repository.