feat(loki): Add Loki project for logs
Initial commit of Loki that collects log data and stores them in an
object store, for display/querying in Grafana.

The installation uses the Loki project
(https://github.com/grafana/loki), manifests generation uses the
`Makefile`, with application using kustomize.

As the upstream project using helm for configuration, most
configuration options are in the values.yaml file. Changes to the
settings, requires rerunning the `make manifests` to create the new
manifests. Commit the results to the metacpan-k8s repository.