fix(web): Configure domains for Anubis
Anubis intelligently picks the host for which it's running and proxies
traffic according to it. This works well, if you're pointing one host at
the service. We're pointing 3, as the hosts use Fastly for caching/DDoS
protection.
Updating the environment variable to include the domains should allow
Anubis to proxy all requests.